extract the mysql connection object
