+++ typescript/src/prices.ts
@@ -1,24 +1,5 @@
 import express from "express";
-import mysql from "mysql2/promise"
-
-export async function createDatabaseConnection(connectionOptions) {
-    let connection = await mysql.createConnection(connectionOptions)
-
-    // make the database connection not too fast so that
-    // we actually feel a slight slowness in the tests
-    return {
-        query(query, params: any[] = []) {
-            return new Promise((resolve) => {
-                setTimeout(resolve, 50)
-            }).then(() => {
-                return connection.execute(query, params)
-            })
-        },
-        close() {
-            return connection.close()
-        }
-    }
-}
+import {createDatabaseConnection} from "./mysql"
 
 async function createApp() {
     const app = express()
